updated portfolio, portfolio_stock models & forms

+++ app/models/portfolio_stock.py
@@ -9,7 +9,7 @@ class PortfolioStock(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     portfolio_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("portfolios.id")), nullable=False)
     stock_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("stocks.id")), nullable=False)
-    quantity = db.Column(db.Integer, nullable=False, default=1)
+    quantity = db.Column(db.Float, nullable=False, default=1)
 
     # Relationships
     portfolio = db.relationship("Portfolio", back_populates="stocks")
